update: 북마크/좋아요 삭제 요청 존재 여부 확인 코드 추가

+++ src/main/java/io/github/bon/wonx/domain/movies/service/BookmarkService.java
@@ -41,7 +41,9 @@ public class BookmarkService {
     // 북마크 삭제
     @Transactional
     public BookmarkDto delete(UUID userId, UUID movieId) {
-        bookmarkRepository.deleteByUserIdAndMovieId(userId, movieId);
+        if (bookmarkRepository.findByUserIdAndMovieId(userId, movieId).isPresent()) {
+            bookmarkRepository.deleteByUserIdAndMovieId(userId, movieId);
+        }
         return BookmarkDto.notBookmarked(userId, movieId);        
     }
 }
